requirements.txt
python-dotenv==1.0.0
requests==2.30.0
urllib3==2.0.2
